+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.8.5] - 2026-09-01
 ### Added
 - MouseScrollHandler
 - GamepadScrollHandler
@@ -114,7 +116,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/UISystem/compare/0.8.4...main
+[Unreleased]: https://github.com/1mbitshorde/UISystem/compare/0.8.5...main
+[0.8.5]: https://github.com/1mbitshorde/UISystem/tree/0.8.5/
 [0.8.4]: https://github.com/1mbitshorde/UISystem/tree/0.8.4/
 [0.8.3]: https://github.com/1mbitshorde/UISystem/tree/0.8.3/
 [0.8.2]: https://github.com/1mbitshorde/UISystem/tree/0.8.2/